Fix Arc3 - end, rotate bitmap test8
역방향 비트맵 회전

+++ intothebreach/intothebreach/mMath.h
@@ -122,7 +122,7 @@ namespace m {
 		}
 
 		float Length() {
-			return sqrtf(x * x + y * y);
+			return sqrt(x * x + y * y);
 		}
 
 		Vector2& Normalize() {
